# [C#] Optimize memory allocations when parsing numbers in .NET5 or greater