Do not use StringBuilder when loading and parsing numbers by using char buffer.
If target is NET5 or greater stackalloc can be used, but due to compatibility reasons we cannot use some Span<> methods.