@jskeet 2024-03-04
Again, I'm not going to have time to get to this any time soon.