@jskeet 2024-03-04
This will conflict with #15794. Do you want me to go ahead getting a final review for that and getting that merged, or is this your preferred approach now?