@jskeet 2024-03-22
Please could update this in light of the current HEAD? I can then try to take a look next week.